requirements.txt
gradio>=4.0
huggingface_hub
gtts
pillow
sqlite3
